+++ ProgrammingInCSharpStore.MyWindowsformApp.Domain/Models/OrderItem.cs
@@ -9,8 +9,8 @@ public class OrderItem: IEntity<long>
     public DateTime CreatedAt { get; set; }
     public bool IsDeleted { get; set; }
     public DateTime DeletedAt { get; set; }
-    public int OrderId { get; set; }
-    public int ProductId { get; set; }
+    public long OrderId { get; set; }
+    public long ProductId { get; set; }
     public int Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public OrderStatus Status { get; set; }
